Forked MS8607 driver to MS5840. Renames+README.
At this point it's still just the MS8607 driver but with a
search-and-replace done to replace "ms8607" with "ms5840" and the README
updated to reflect the fact that this is intended for a different
driver. So it probably won't work for the ms5840 in this state, and
might not even compile. The more detailed work has yet to come.